Add a cursor
Unlike the iterator, the cursor's next method provides the summary at
the start of the element it yields. This summary is borrowed from the
cursor, which isn't possible to implement for the iterator trait.

The cursor can also be used to build a prefix tree up to a specific
dimension. After building the tree, the cursor is parked on the element
following this prefix.