Batch atlas updates
We are observing a fixed cost when uploading new data to the texture, so
when adding a large number of glyphs to the atlas (i.e., during editor
startup) it is more efficient to upload all of them at once.

Because `getImageData` is slow, the simplest solution for now is to
re-upload the entire texture.

Co-authored-by: Nathan Sobo <nathan@github.com>