Use `cfg` instead of features to target Node.js versions
Cargo features aren't best suited for this sort of conditional
compilation as the package consuming `napi` shouldn't need to be
involved in the Node.js v8/v9 compatibility details.

Therefore it's more straightforward to use the `build.rs` script
within `napi` to toggle the right `#[cfg]` attribute to
conditionally compile the correct module depending on the detected
Node.js version. This keeps *all* of those gubbins inside of `napi`
and prevents them from causing headaches for packages using it.