Cache mapping between offsets and points
When moving cursors, we heavily rely on `Buffer::len_for_row` to
understand when such cursors should wrap. This method is implemented by
traversing the tree twice: the first traversal gets the offset of the
beginning of the requested line, whereas the second traversal gets the
offset of the line next to the requested one. Then, the two offsets are
subtracted to obtain the length of the line.

With this commit we start caching the mapping between offsets and
points. This allows to skip the two traversals and retrieve the length
of a line in constant time if the line length had been computed before
and no edit has occurred since then.