Introduce View React component
This React component will interact with the view registry to render a
the component associated with the supplied view id. When the
corresponding props change, this component will re-render its subtree.

`View` also exposes a `dispatch` function to its underlying component.
This is a convenient wrapper that automatically routes actions to the
appropriate view on the server.