Add action dispatching to ViewRegistry
To keep ViewRegistry a pure JavaScript component, calling
`dispatchAction` will simply check if the view exists in the local state
and then call `onAction` if it was supplied.

By designing things this way we can:

* Test other components that rely on ViewRegistry more easily, without
necessarily introducing a fake client
* Keep ViewRegistry's logic separated from sending messages. This way we
can have a single place that interacts with XrayClient both when
messages are received and when messages are sent.