Implement file system tree scanning
This commit adds an `fs` module to xray_core, which exposes the Tree
trait along with Dir and File structures that are designed for
concurrent access.

In `xray_server`, we implement the Tree trait and spawn a thread that
populates the tree in the background. The current state of the tree can
be read in the foreground as it is being populated.