Reduce memory manipulation in FuzzySearch::process
This allows us to scan the entire Atom folder within 60ms when searching
for a query like "texteditorcomponent".

As a consequence of the above performance improvements, with this commit
we can afford scanning more entries before reporting search results, so
we are increasing `DEFAULT_ENTRY_COUNT_PER_POLL` constant to 100k.