Fix warnings
These are all related to not using the results of a local channel after
sending some message on them. We are fixing such warnings by unwrapping
the return value of `send`. This should be a safe operation because in
all cases both the receiving end and the transmitting end of the channel
have the same lifetime.