WIP – Go back to having Service implement poll_update
What prompted this change? We were implementing the update stream on the
App service, and since app is simple, we can just call `state()` on
every update. Inside a call to `poll_update` that is passed a reference
to the connection, calling `state` would have been trivial. Calling it
from a static-lifetime stream that's running separately from the service
object would introduce a ton of complexity.

So, we think we should switch back to having Service implement
poll_update directly rather than yielding a stream. Previously, we were
thinking that model objects would implement the service trait
themselves. Now we're thinking it might be a better trade-off to have
separate service objects similar to how we deal with views. So we'd have
a Workspace, a WorkspaceView, and a WorkspaceService.

Service objects can now maintain a reference to a shared model object,
while also maintaining a notify cell observer or any other local state
that they need to handle calls to poll_update. The big advantage here is
that we have access to the connection during poll_update so that we can
add child services.

Co-authored-by: Max Brunsfeld <maxbrunsfeld@github.com>