Add client::FullUpdateService
This helper can wrap client::Service<T> where T is a server::Service
that does not provide incremental updates (i.e., T::State == T::Update).

client::FullUpdateService is very similar to client::Service, but
differs from it in that it will keep the state up to date as updates are
integrated from the server.

Co-authored-by: Nathan Sobo <nathan@github.com>