WIP – Start on project replication
Sketched in a service and a remote project type. There's some
duplication around searching projects but there's not actually that much
logic there. It's mostly forwarding data. I tried to extract a `trees`
method but the refcell on the local project that's required to share its
state with the service makes it complicated to give the method a
consistent interface without wrapping the remote project's state in a
refcell also.

When constructing the initial state, the ProjectService should eagerly
create child services for all of the trees and track them. That will
allow the RemoteProject to obtain a service for each RemoteTree in the
project.

Then we'll need to implement request/response for opening buffers.