Avoid redundant notifications in xray_wasm Executor
When we call notify on a notify handle, we add its future to acollection
of pending futures. If this is the *first* notificationrequest since the
last time we polled, we call into JS to request a pollon the next tick.
This avoid redundant polls of the same future whenmultiple notifications
are triggered, which is incorrect behavior. Italso minimizes
interactions with JS and the resultant JS heapallocations.

/cc @as-cii

Co-Authored-By: Jason Rudolph <jason@jasonrudolph.com>