Teach script/build to establish necessary wasm-bindgen version
We need the right version of wasm-bindgen in order to build xray_wasm.
Having the wrong version of wasm-bindgen leads to the issue described in
https://github.com/atom/xray/issues/72. The changes in
https://github.com/atom/xray/issues/71 improved this situation for our
Travis CI builds, but we're running into a similar issue in our local
dev builds. This commit moves the wasm-bindgen setup out of .travis.yml
and into xray_wasm/script/build so that we can benefit from it both for
local dev and for CI.