Optimize Buffer::merge_selections
Instead of using `Vec::remove` when merging adjacent selections, we now
construct a new Vec with the merged selections. In the worst case, this
corresponds to a single movement of memory from a vector to another and
a single malloc. Using `Vec::remove` repeatedly, on the other hand, was
causing the vector to be constantly shifted around, which starts
becoming slow when the number of selections increases dramatically. In
my tests, this change does not introduce slowdowns when dealing with a
small number of selections.