Add Buffer::insert_selections and use it in add_selection_{above,below}
This method is very similar to Buffer::mutate_selections in that it
allows to manipulate selection sets by supplying an id. The only
difference with the latter is that it gives users of this API a readonly
slice of the existing selections and allows to return a (not necessarily
ordered) vector of new selections to insert.