Start on Builder
The builder will be created with an existing tree and a path to a
directory within that tree (the root directory is allowed). Then the
builder's owner will call push_dir, push_file, and pop_dir, on the
builder, which will compare the pushed entry with the previous tree and
insertions and deletions as needed.

Currently, the builder only handles super simple insertions into empty
directories.