WIP: Start on fs2 module with alternative approach to representing tree
The idea here is not to represent the file system as a tree of paths.
Instead we will populate the btree with heterogenous items. Items are
grouped together by their file_id, and each item is either a ParentRef
or a ChildRef. Files only have one or more ParentRefs that indicate the
id of directories that contain the file. Directories have child refs
that list the identifiers of the directory's entries.