Order ChildRefs by timestamp and replica_id
The `deletions` vector will now only be used when an actual removal 
occurred and never to indicate that a tie occurred. Also, this way we 
will be able to add a ChildRef for the same `ref_id` more easily, which 
could happen when moving a directory out of a parent dir and then moving 
it back with the same name. In that case the timestamp will be 
positioned earlier in the tree and will win over the previous entry.