Handle deleted child refs that conflict with existing names
In rare circumstances, we may end up inserting a deleted child ref that 
conflicts with an existing name. Previously, that would cause an 
invariant violation in which a deleted child ref would be followed by a 
visible child ref with the same name. This commit solves that by 
synthesizing a move of the existing directory entry to the same name so 
that it has a higher timestamp.