WIP: Start on TreeDiffCursor
It will yield items which describe a diff between two trees, indicating 
which entries are the same, inserted, or removed. The current commit has 
no tests, and we will also need to include an optional src_path in 
inserted entries and dst_path in removed entries to indicate moves. I 
decided to yield everything in a single item to avoid redundant 
comparisons and calls down into the underlying tree cursor. It makes me 
wonder if we want the tree cursor to have a similar batched interface 
where we return a single item value with multiple fields of information.

Co-Authored-By: Nathan Sobo <nathan@github.com>