Batch conflict resolution
Since operations are produced in batches based on snapshots of the file 
system, it's possible for the final outcome of a set of operations to 
produce a valid tree even if some of those operations would produce 
conflicts when applied individually. To avoid false positives, we need 
to wait until all operations have been applied and then fix conflicts.

Co-Authored-By: Antonio Scandurra <as-cii@github.com>