Duplicate dir if a remote move is concurrent with a mistaken removal
A directory is mistakenly considered as removed when it is moved 
backwards in the lexicographic ordering of the tree while traversing it.

On a subsequent scan we may discover that this directory still exists, 
but that we have associated it with a different inode when handling a 
concurrent remote move operation. This commit causes us to treat the 
directory as if it were new in this scenario. A more correct solution 
would be to merge the two directories back together and preserve 
identity, but that would introduce a lot of complexity for a scenario 
that we think is extremely unlikely in practice.

Co-Authored-By: Nathan Sobo <nathan@github.com>