Don't reflect failure to rename dirs with temp names in the new tree
If we can't rename a directory that has a temporary name, it means that 
the file system has changed and the old tree is stale. 

Previously, we tried to update the index to reflect the state of the 
file system but this was not guaranteed to succeed because a concurrent 
file system change may have introduced an additional name conflict. Now 
we will just wait for the next scan to synchronize the index.

Co-Authored-By: Nathan Sobo <nathan@github.com>