WIP: Start on WorkingCopy
I think the best place for us to start is to basically replicate the 
previous functionality of the Timeline in the form of the WorkingCopy. 
Eventually, the WorkingCopy will need a more sophisticated API to deal 
with committing etc, but for now we can just focus on getting basic 
replication working starting from a particular set of base paths and 
applying operations to an empty patch. This will force us to think about 
how to store the full set of paths in a separate structure from the 
sparse CRDT and how to interrelate the two. I think maintaining a 
bi-directional mappings between locally assigned ids in the index and 
remote ids in the CRDT will be a key part of this. I sketched in a test 
that should drive out some of this design.