Start over from WorkTree
This new module will still be sparse in the sense that it will emit only 
operations based on a common state that does not need to be shared over 
the wire. At the same time, however, it will also be dense because it 
requires inserting all the common paths into the data structure.