Give buffer base text an id with a sequence number of zero
Since version vectors include sequence number 0 by default for any site, 
this means that the base text is always considered to be included in any 
version vector without special treatment. This simplifies the 
construction of the version vector in the WorkTree because we don't have 
to include sequence number 1 for replica 0 in order to represent the 
base text like we did previously.