Use a request/response style for calls from JS into Rust
Rather than binding individual methods, we take a more server-like 
approach and just have one request method. This takes a JS object and 
deserializes it into a request, then spits out a response. Operations 
and file ids are represented as base64 encoded strings to simplify 
handling on the JS side.