Keep text of opened buffers in sync when switching to a new epoch
This commit will compute all the changes that occurred between buffers 
in the old epoch  and in the new epoch and report them to the observer. 

This implies that any outstanding change that occurred between the 
creation of the epoch and the moment we actually switch to it gets lost. 
This limitation is unfortunate but it allows us to ship a working 
implementation of epochs.

In the future we may want to revisit some aspects of this commit, such 
as:

1. The limitation described above. By using a version vector and a 3-way 
diff we should be able to carry outstanding changes forward to the next 
epoch (so long as they don't introduce conflicts).
2. Consider using a different implementation of the diff algorithm. The 
difference.rs crate is convenient to use, but unnecessarily allocates 
strings and has an average time complexity of O(m * n). Considering 
strings will often be similar, it would be nice to use an algorithm like 
Myers diff which scales with the number of edits.