Fix reporting changes to observer after applying ops
Previously, we were relying on the epoch version to report changes to
the observer. Unfortunately, such version doesn't fully reflect the
state of the buffer, as some operations may have been deferred. This
implies that the epoch sometimes will indicate that some changes have
been incorporated when, in reality, the observer has never seen them.

For now, I am circumventing this problem by relying on each buffer's
version to report changes, as opposed to using the epoch's version. This
is okay for now and makes the failing seed work. As soon as we start
dealing with partial commits, however, we will be required to broadcast
a version vector which, potentially, does not reflect the state of the
local buffers. This has implications in terms of the diffs we will
compute and must be addressed before we can implement partial commits
and soft resets properly.

There could be two solutions in that case:

1. Broadcast the epoch's version as well as each buffer's seen version.
This should work correctly, although it implies sending extra data to
clients. Considering the number of open buffers is usually small, this
might be a good solution that doesn't require a huge time investment.
2. Defer buffer operations at the epoch layer and don't process further
operations coming from a particular site until the first deferred
operation can be unblocked. Of the two approaches, this is the most
elegant but requires more work due to the fact that we simply push
buffer operations to a vector when such buffer hasn't been opened yet.
So, in order to implement this approach, we first need to create Buffer
instances even when we don't have the base text. This is something we
want to do anyway in order to improve the correctness of
`Epoch::is_modified_file`.

Finally, this commit makes our randomized test suite pass, so the next
step will be to integrate WorkTree with the JavaScript layer.