Don't create multiple FlatBuffer tables at the same time
Specifically, we were serializing buffer operations _after_ starting a 
Vector in epoch::Operation::to_flatbuf. This was throwing an error 
because Vector is a table, and serializing buffer operations creates 
strings, which are also tables. This is forbidden in FlatBuffers.

This commit serializes all the buffer  operations first and then creates 
a vector out of those.