Replace observer's text_changed and selections_changed with changed
This new method will include both the changes and the selection ranges 
after applying the operations or switching epoch. This was driven by 
wasm-bindgen's limitation of calling re-entrantly into the Rust layer 
(e.g. Rust -> JS -> Rust is not allowed when that would cause a borrow 
failure).

Since we can't really allow the observer to call back into Rust, we are 
removing the code that was dropping the RefCell borrows.