Remove time::Local fields from Operation::UpdateSelections
This commit makes `set_id` a `time::Lamport` as well as removing the 
`local_timestamp` field.

Selection updates always include the entire state of selections and do 
not depend on previous updates of the same selection set. As such, users 
of memo may not want to store all of them and memo should be capable of 
supporting that model. Using time::Local in the operation payload would 
create "holes" in the version vector that could be problematic or, at 
least, awkward.

By using Lamport timestamps we can still track causality of operations 
and not introduce any hole in the operation history.

Co-Authored-By: Jason Rudolph <jason@jasonrudolph.com>