Introduce replica locations API
This commit introduces the new `set_active_location` and 
`replica_locations` methods on WorkTree, which respectively allow to set 
the active buffer and retrieve the location of all the peers in the 
current epoch. In the future, I am imagining we will tweak this API to 
also include a SelectionSetId.